+++ src/Fable.ReactTestingLibrary/AssemblyInfo.fs
@@ -6,8 +6,8 @@ open System.Runtime.CompilerServices
 [<assembly: AssemblyTitleAttribute("Fable.ReactTestingLibrary")>]
 [<assembly: AssemblyProductAttribute("Fable.Jester")>]
 [<assembly: AssemblyDescriptionAttribute("Fable bindings for jest and friends")>]
-[<assembly: AssemblyVersionAttribute("0.5.0")>]
-[<assembly: AssemblyFileVersionAttribute("0.5.0")>]
+[<assembly: AssemblyVersionAttribute("0.6.0")>]
+[<assembly: AssemblyFileVersionAttribute("0.6.0")>]
 [<assembly: AssemblyConfigurationAttribute("Release")>]
 [<assembly: InternalsVisibleToAttribute("Fable.ReactTestingLibrary.Tests")>]
 do ()
@@ -16,7 +16,7 @@ module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "Fable.ReactTestingLibrary"
     let [<Literal>] AssemblyProduct = "Fable.Jester"
     let [<Literal>] AssemblyDescription = "Fable bindings for jest and friends"
-    let [<Literal>] AssemblyVersion = "0.5.0"
-    let [<Literal>] AssemblyFileVersion = "0.5.0"
+    let [<Literal>] AssemblyVersion = "0.6.0"
+    let [<Literal>] AssemblyFileVersion = "0.6.0"
     let [<Literal>] AssemblyConfiguration = "Release"
     let [<Literal>] InternalsVisibleTo = "Fable.ReactTestingLibrary.Tests"